chage createAllOfTablesWithDataSourceORM parameter

+++ src/main/java/com/revature/banking/orm/utils/CrudORM.java
@@ -50,19 +50,20 @@ public class CrudORM {
         return false;
     }
 
-    public <T> void createTable(Class<T> newData) {
-
+    public <T> boolean createTable(Class<T> newData) {
         try (Connection conn = ConnectionFactory.getInstance().getConnection()) {
             DataSourceORM tableAnnotation = newData.getAnnotation(DataSourceORM.class);
+            if( tableAnnotation == null){
+                return false;
+            }
             String tableName = tableAnnotation.TableName();
             String schema = tableAnnotation.Schema();
-            List<Field> fields = ReflectionORM.getFieldNamesAndValues(newData);
+            if (existsTable(conn, schema, tableName))
+                return true;
 
-            if (existsTable(conn, schema, tableName)) return;
-
+            List<Field> fields = ReflectionORM.getFieldNames(newData);
             StringBuilder sql = new StringBuilder();
             sql.append(" CREATE TABLE ").append(schema).append(".").append(tableName).append(" ( ");
-
             for (Field f : fields) {
                 sql.append(f.getName()).append(" ").append(convertTypeFromJaveToPostgresql(f.getType())).append(" ");
                 ColumnInORM columnInORM = f.getAnnotation(ColumnInORM.class);
@@ -115,52 +116,50 @@ public class CrudORM {
             }
             sql.setLength(sql.length() - 2);
             sql.append(" ); ");
-
             logger.info(sql);
-
             PreparedStatement pstmt = conn.prepareStatement(sql.toString());
             int rowsInserted = pstmt.executeUpdate();
             if (rowsInserted != 0) {
                 logger.info("{} -- table created", tableName);
+                return true;
             }
         } catch (SQLException e) {
             e.printStackTrace();
         }
+        return false;
     }
 
-    public <T> List<T> readTable(T newData, Map<String, Map<String, String>> whereOderBy, Class<T> cls) {
+    public <T> List<T> readTable(T newData, Map<String, Map<String, String>> whereOrderBy, Class<T> cls) {
         List<T> ormList = new LinkedList<>();
         try (Connection conn = ConnectionFactory.getInstance().getConnection()) {
             Class<?> aClass = newData.getClass();
             DataSourceORM tableAnnotation = aClass.getAnnotation(DataSourceORM.class);
+            if( tableAnnotation == null){
+                return null;
+            }
             String tableName = tableAnnotation.TableName();
             String schema = tableAnnotation.Schema();
-            List<Field> fields = ReflectionORM.getFieldNamesAndValues(aClass);
+            List<Field> fields = ReflectionORM.getFieldNames(aClass);
 
             StringBuilder sql = new StringBuilder();
             sql.append(" select * from ").append(schema).append(".").append(tableName).append(" ");
-
-            if (whereOderBy.containsKey("where") && whereOderBy.get("where").size() > 0) {
+            if (whereOrderBy.containsKey("where") && whereOrderBy.get("where").size() > 0) {
                 sql.append("where ");
-                for (Map.Entry<String, String> where : whereOderBy.get("where").entrySet()) {
+                for (Map.Entry<String, String> where : whereOrderBy.get("where").entrySet()) {
                     sql.append(where.getKey()).append(" = '").append(where.getValue()).append("', ");
                 }
                 sql.setLength(sql.length() - 2);
             }
-            if (whereOderBy.containsKey("oderBy") && whereOderBy.get("oderBy").size() > 0) {
-                for (Map.Entry<String, String> cwo : whereOderBy.get("oderBy").entrySet()) {
+            if (whereOrderBy.containsKey("oderBy") && whereOrderBy.get("oderBy").size() > 0) {
+                for (Map.Entry<String, String> cwo : whereOrderBy.get("oderBy").entrySet()) {
                     sql.append(cwo.getKey()).append(" ").append(cwo.getValue()).append(", ");
                 }
                 sql.setLength(sql.length() - 2);
             }
             sql.append(" ; ");
-
             logger.info(sql);
-
             PreparedStatement pstmt = conn.prepareStatement(sql.toString());
-
             ResultSet rs = pstmt.executeQuery();
-
             String CapitalFirst = "";
             while (rs.next()) {
                 T appEntityORM = cls.newInstance();
@@ -168,64 +167,64 @@ public class CrudORM {
                 for (Field f : fields) {
                     CapitalFirst = f.getName().substring(0, 1).toUpperCase() + f.getName().substring(1);
 
-                    Method sumInstanceMethod = aClass.getMethod("set" + CapitalFirst, String.class);
+                    Method aMethod = aClass.getMethod("set" + CapitalFirst, String.class);
 
                     if (f.getType().isAssignableFrom(LocalDateTime.class)) {
-                        sumInstanceMethod.invoke(appEntityORM, rs.getTimestamp(f.getName()).toString().substring(0, 19));
+                        aMethod.invoke(appEntityORM, rs.getTimestamp(f.getName()).toString().substring(0, 19));
                     } else {
-                        sumInstanceMethod.invoke(appEntityORM, rs.getString(f.getName()));
+                        aMethod.invoke(appEntityORM, rs.getString(f.getName()));
                     }
                 }
-
                 ormList.add(appEntityORM);
             }
-        } catch (SQLException | InstantiationException | IllegalAccessException | NoSuchMethodException | InvocationTargetException e) {
+        } catch (SQLException | InstantiationException | IllegalAccessException | NoSuchMethodException |
+                InvocationTargetException e) {
             e.printStackTrace();
         }
         return ormList;
     }
 
-    public <T> T updateTable(T newData, Map<String, Map<String, String>> colsWhereOderBy) {
+    public <T> T updateTable(T newData, Map<String, Map<String, String>> colsWhereOrderBy) {
         try (Connection conn = ConnectionFactory.getInstance().getConnection()) {
 
             Class<?> aClass = newData.getClass();
             DataSourceORM tableAnnotation = aClass.getAnnotation(DataSourceORM.class);
+            if( tableAnnotation == null){
+                return null;
+            }
             String tableName = tableAnnotation.TableName();
             String schema = tableAnnotation.Schema();
-            List<Field> fields = ReflectionORM.getFieldNamesAndValues(aClass);
+            List<Field> fields = ReflectionORM.getFieldNames(aClass);
 
             StringBuilder sql = new StringBuilder();
             sql.append("update ").append(schema).append(".").append(tableName).append(" set ");
             // ------------ column
-            if (colsWhereOderBy.containsKey("cols") && colsWhereOderBy.get("cols").size() > 0) {
-                for (Map.Entry<String, String> col : colsWhereOderBy.get("cols").entrySet()) {
+            if (colsWhereOrderBy.containsKey("cols") && colsWhereOrderBy.get("cols").size() > 0) {
+                for (Map.Entry<String, String> col : colsWhereOrderBy.get("cols").entrySet()) {
 
                     String CapitalFirst = col.getKey().substring(0, 1).toUpperCase() + col.getKey().substring(1);
-                    Method sumInstanceMethod = aClass.getMethod("get" + CapitalFirst);
+                    Method aMethod = aClass.getMethod("get" + CapitalFirst);
 
-                    if (sumInstanceMethod.getReturnType().isAssignableFrom(LocalDateTime.class) ||
-                            sumInstanceMethod.getReturnType().isAssignableFrom(Double.class)) {
-                        sql.append(col.getKey()).append(" = ").append(sumInstanceMethod.invoke(newData)).append(", ");
+                    if (aMethod.getReturnType().isAssignableFrom(LocalDateTime.class) ||
+                            aMethod.getReturnType().isAssignableFrom(Double.class)) {
+                        sql.append(col.getKey()).append(" = ").append(aMethod.invoke(newData)).append(", ");
                     } else {
-                        sql.append(col.getKey()).append(" = '").append(sumInstanceMethod.invoke(newData)).append("', ");
+                        sql.append(col.getKey()).append(" = '").append(aMethod.invoke(newData)).append("', ");
                     }
                 }
                 sql.setLength(sql.length() - 2);
             }
             // ------------ where
-            if (colsWhereOderBy.containsKey("where") && colsWhereOderBy.get("where").size() > 0) {
+            if (colsWhereOrderBy.containsKey("where") && colsWhereOrderBy.get("where").size() > 0) {
                 sql.append(" where ");
-                for (Map.Entry<String, String> wh : colsWhereOderBy.get("where").entrySet()) {
+                for (Map.Entry<String, String> wh : colsWhereOrderBy.get("where").entrySet()) {
                     sql.append(wh.getKey()).append(" = '").append(wh.getValue()).append("', ");
                 }
                 sql.setLength(sql.length() - 2);
             }
             sql.append(";");
-
             PreparedStatement pstmt = conn.prepareStatement(sql.toString());
-
             logger.info(sql);
-
             int rowsInserted = pstmt.executeUpdate();
             if (rowsInserted != 0) {
                 return newData;
@@ -242,10 +241,12 @@ public class CrudORM {
 
             Class<?> aClass = newData.getClass();
             DataSourceORM tableAnnotation = aClass.getAnnotation(DataSourceORM.class);
+            if( tableAnnotation == null){
+                return null;
+            }
             String tableName = tableAnnotation.TableName();
             String schema = tableAnnotation.Schema();
-            List<Field> fields = ReflectionORM.getFieldNamesAndValues(aClass);
-
+            List<Field> fields = ReflectionORM.getFieldNames(aClass);
             // ------------
             StringBuilder sql = new StringBuilder("insert into " + schema + "." + tableName + "( ");
             // ------------ column
@@ -255,7 +256,6 @@ public class CrudORM {
                 sql.append(f.getName()).append(", ");
             }
             sql = new StringBuilder(sql.substring(0, sql.length() - 2));
-
             // ------------ values placeholders
             sql.append(") values (");
             for (Field f : fields) {
@@ -265,64 +265,56 @@ public class CrudORM {
             }
             sql = new StringBuilder(sql.substring(0, sql.length() - 2));
             sql.append(")");
-
             PreparedStatement pstmt = conn.prepareStatement(sql.toString());
             // ------------ values real
             for (int i = 0; i < fields.size(); i++) {
                 if (fields.get(i).getAnnotation(NotIntoDabase.class) != null)
                     continue;
-
                 String fn = fields.get(i).getName();
                 String cap = fn.substring(0, 1).toUpperCase() + fn.substring(1);
-
-                Method sumInstanceMethod = aClass.getMethod("get" + cap);
-
+                Method aMethod = aClass.getMethod("get" + cap);
                 if (fields.get(i).getType().isAssignableFrom(Double.class)) {
-                    pstmt.setDouble(i + 1, (Double) sumInstanceMethod.invoke(newData));
+                    pstmt.setDouble(i + 1, (Double) aMethod.invoke(newData));
                 } else {
-                    pstmt.setString(i + 1, (String) sumInstanceMethod.invoke(newData));
+                    pstmt.setString(i + 1, (String) aMethod.invoke(newData));
                 }
             }
-
             logger.info(pstmt.toString());
             System.out.println(pstmt.toString());
-
             int rowsInserted = pstmt.executeUpdate();
             if (rowsInserted != 0) {
                 return newData;
             }
-
         } catch (SQLException | NoSuchMethodException | InvocationTargetException | IllegalAccessException e) {
             e.printStackTrace();
         }
         return null;
     }
 
-    public <T> T deletTable(T newData, Map<String, Map<String, String>> colsWhereOderBy) {
+    public <T> T deletTable(T newData, Map<String, Map<String, String>> colsWhereOrderBy) {
         try (Connection conn = ConnectionFactory.getInstance().getConnection()) {
-
             Class<?> aClass = newData.getClass();
             DataSourceORM tableAnnotation = aClass.getAnnotation(DataSourceORM.class);
+            if( tableAnnotation == null){
+                return null;
+            }
             String tableName = tableAnnotation.TableName();
             String schema = tableAnnotation.Schema();
-            List<Field> fields = ReflectionORM.getFieldNamesAndValues(aClass);
+            List<Field> fields = ReflectionORM.getFieldNames(aClass);
 
             StringBuilder sql = new StringBuilder();
             sql.append("delete from ").append(schema).append(".").append(tableName);
             // ------------ where
-            if (colsWhereOderBy.containsKey("where") && colsWhereOderBy.get("where").size() > 0) {
+            if (colsWhereOrderBy.containsKey("where") && colsWhereOrderBy.get("where").size() > 0) {
                 sql.append(" where ");
-                for (Map.Entry<String, String> wh : colsWhereOderBy.get("where").entrySet()) {
+                for (Map.Entry<String, String> wh : colsWhereOrderBy.get("where").entrySet()) {
                     sql.append(wh.getKey()).append(" = '").append(wh.getValue()).append("', ");
                 }
                 sql.setLength(sql.length() - 2);
             }
             sql.append(";");
-
             PreparedStatement pstmt = conn.prepareStatement(sql.toString());
-
             logger.info(sql);
-
             int rowsInserted = pstmt.executeUpdate();
             if (rowsInserted != 0) {
                 return newData;
@@ -333,8 +325,8 @@ public class CrudORM {
         return null;
     }
 
-    public <T> void createAllOfTablesWithDataSourceORM(Class<T> aClass) throws SQLException, URISyntaxException {
-        System.out.println("Package Name in InitORM.packageNameInOrm = " + InitORM.packageNameInOrm);
+    public <T> void createAllOfTablesWithDataSourceORM(T newData) throws SQLException, URISyntaxException {
+        Class<?> aClass = newData.getClass();
         List<String> listClassesNames = ReflectionORM.getClassNamesInPackage(aClass, InitORM.packageNameInOrm);
         List<Class<?>> classList = listClassesNames.stream()
                 .map(ClassNameToClassMapper.getInstance())
